move code from ruuvi-hass.py to sensor.py
